Reverting changes on start(...) function... there is a bug on libcds.

+++ tgs.cpp
@@ -67,8 +67,8 @@ size_t start(BitSequence *b, uint i) {
 //	return b->select1(i+1) - i; 	//in the paper this operation is
                                   //start(i) = select1(b, i) - i + 1
 
-	size_t ret = b->select1(i+1);
-  if (ret == (size_t)(-1)) {
+	uint ret = b->select1(i+1);
+  if (ret == (uint)-1) {
     return b->getLength() - i;
   }
   else {
